Add Not-For-Us checking to cruft_report
This patch reports all binary packages that are out of date on
an architecture where the buildd admin has marked it as Not-For-Us.

It needs a copy of http://buildd.debian.org/stats/*-all.txt somewhere.
This somewhere is configurable by a configuration/command-line flag.

The report, named "nfu", is added to the "full" set of reports.