dak/process_new.py: handle uploads to (oldstable)-proposed-updates differently
Uploads targetting proposed-updates used to hit NEW first and were going
to proposed-updates without further review by SRM afterwards.  This patch
fixes this behaviour by sending the upload to p-u-new after NEW processing.

If an upload is sourceful it is moved right away.  If it is not its
corresponding source is searched in p-u holding, new, accepted and
in the archive and moved appropriately into the right queue.