md5sum check, passive mode
Only check md5sums after transfering stuff to the target if we are
configured to do this. We can't do this via ftp, as the files on
ftp.upload.debian.org aren't readable for us (vsftpd runs as a user
not allowed to read them).

Also, no need to check them - if the files are really broken then the
next queued will detect that and mail the maintainer anyways. So it
wont get broken files near the archive.

Signed-off-by: Joerg Jaspert <joerg@debian.org>