Changed process_unchecked.py to handle DM's (slightly) more sanely.
Previously, process_unchecked only checked unstable to determine DM upload rights.

This caused issues with experimental since a package could have DM-Allowed-Upload in experimental
but not in stable causing a REJECT.

The applied code change causes dak to find the highest version of a package, and use its uploaders/DM-Allowed-Uploader
flag to determine upload rights to that package. This also resolves a corner case usage where a package only exists
in testing/stable/oldstable, and upload rights to {t-|o-}p-u.

Note: In addition to the previous ALTER TABLE statememnt, the following SQL query is necessary to
      finish the schema updates. This does not apply to installations not using DMs.

      UPDATE source SET dm_upload_allowed = 't' WHERE id = (SELECT source FROM src_uploaders);

Signed-off-by: Michael Casadevall <sonicmctails@gmail.com>