update dinstall.html for 4 daily dinstall runs
dinstall.html contains a javascript countdown timer used to estimate when the
next dinstall run will occur.  It was hardcoded for 2 daily dinstall runs.
This patch hard codes it to 4 daily runs.

There previously was a link to the announcement of the switch to 2 daily
dinstall runs.  I was not able to locate a similar announcement to the switch
to 4 daily runs, so I simply removed the link to the announcement.

Signed-off-by: Mike O'Connor <stew@vireo.org>