generate_releases
depending on what kind of file we have we generate the checksums in
different ways. For the uncompressed sum we continue reading in all the
content and hand that as a variable to apt_pkg.*sum. Turns out this does
not work for the already compressed files, there we now go and just hand
the opened file_handle to apt_pkg.*sum.