add a class providing access methods to a projectb database
This class provides an alternative to daklib/database.py for accessing a
projectb database.  This benefits to this alternative are that it uses the
psycopg2 module instead of the pg module, and I object oriented.  My
motivation for creating this class is to keep the new code for generating
Contents.gz files free from the pg module, and therefore, the only functions
I've moved into this class so far are those needed for generating the
Contents.gz files.

Signed-off-by: Mike O'Connor <stew@vireo.org>