Check for the existance of the function before trying to drop it.
At least on my machine, ignoring the error if the function doesn't exist
doesn't work since psycopg2 bombs.  So, we attempt to at least check if
there are any functions with a similar name before trying to drop it.
This way it should work better on an empty database where the function
has never existed.

Signed-off-by: Allan Lyons <Allan_Lyons@wycliffe.ca>