Add start of a metapackage to just get dak running.
Commit 91dbbbfe193962ede641168e2c446ad807f9217a suggested that the
ftpmaster-lenny metapackage could be split in two parts and that a
ftpmaster-common metapackage could be created to document the
requirements to just get dak running. Other tools normally required
for the ftpmaster to actually do their work are in "ftpmaster-$release"

The reason for splitting out the dak requirements separately are in the
hope that someone will eventually do a complete package for dak.  This
should get them started on the requirements.

Postgresql-plperl is Recommends: rather than Depends: on the theory that
the database could be installed on a different machine.

This is not yet checked for completeness, but should give a start
to what dak actually requires to run.

Signed-off-by: Joerg Jaspert <joerg@debian.org>

Signed-off-by: Allan Lyons <Allan_Lyons@wycliffe.ca>