process-new
no longer check the daily lock when we start looking at a package, but
when we try to do modifying actions.
No reason to not allow people to do Check/Reject/Note, those are actions
we can do whenever. Its just modifying overrides which *have* to be
avoided at certain times.

Signed-off-by: Joerg Jaspert <joerg@debian.org>