p-n
look process-new with an own lockfile.
This wont make a difference for dinstall, same lock time there.
But it will make a big difference whenever cron.unchecked runs (which is far
more often), as it now will lock for a few seconds and not for the
sometimes unbearable long times process-unchecked and the w-b trigger
needs.

Signed-off-by: Joerg Jaspert <joerg@debian.org>