Send the "transitions completed" mail to debian-devel instead of the RT.
As per http://lists.debian.org/debian-devel/2009/05/msg00673.html, notice
of transition blocks being lifted is going to be sent to debian-devel and
not to the release team.

This commit also reworks the wording of the body of the message (the
important bit being that the block is lifted, rather than the transitions
were removed from the transitions.yaml file), and it includes the names
of the finished transitions in the Subject of the mail.

Signed-off-by: Adeodato Simó <dato@net.com.org.es>