dinstall
background two function calls.
once for the initial ssh to merkel and the other the dak-dev cluster
projectb reload on ries.
The first does not need to be foreground, we do not care if we are already
continuing in dinstall, its only informative for the qa user on merkel.
The second also doesnt need to run alone. The next thing after it is the 1hour
apt-fuckthearchive call. So the dak-dev reload will well be done, with its 15 minutes
runtime, before the next function is also done. And there is no harm done
in having the two run at the same time.

Signed-off-by: Joerg Jaspert <joerg@debian.org>