dqeued
it is really helpful if we upload to ftp-master.d.o instead of ftp.upload.debian.org.
Especially in the case ftp.u.d.o points back to us, as then we think its already there
and go "HEY, FINE, DELETE". Which isnt what the DDs expect, for some reason.
Thanks to sgran/weasel for this fix.

Signed-off-by: Joerg Jaspert <joerg@debian.org>