check_dsc_against_db: fix a corner-case in .orig. handling
If the .orig.tar was contained in the .changes file but already
present in the pool, we did delete it from the file list, but
did not record the files id in the desc file list.

In the old code this did not cause any errors since process-accepted
called check_dsc_against_db with the file list loaded from the .dak
file which was already cleansed of the .orig. So it took the other
code branch (.orig. present in the pool but not in changes) which
did record the files id.

Now we only call check_dsc_against_db once in process-upload so
we need to record the files id in this case, too.

Signed-off-by: Frank Lichtenheld <djpig@debian.org>