queue, check_source
only bail out if we have a non-source upload.
do not bail out if self.pkg.orig_files is empty.
We either have a bug filling orig_files where we only set this when the files
already have been in the archive, or we intended to use orig_files to show
us files already in the archive. Currently it looks like the latter, and then
it is wrong to check on that in check_source, as that would mean we skip one
check (and also extraction of .debinfo information for the BTS!) for every
upload that comes with totally new files. (Imagine a native package).

Signed-off-by: Joerg Jaspert <joerg@debian.org>