daklib/utils.py: use socket.getfqdn() to determine the hostname
To: debian-dak@lists.debian.org
Cc: Philipp Kern <pkern@debian.org>
Date: Tue, 23 Feb 2010 11:35:20 +0100
Resent-From: debian-dak@lists.debian.org

The current socket.gethostbyaddr(socket.gethostname()) fails horribly on
our IPv6 enabled machine.  What we really want is the FQDN of the host so
let's use socket's getfqdn().  Of course that needs proper setup of the
hostname in the system, but I think we can rely on that working.

Signed-off-by: Philipp Kern <pkern@debian.org>