Add functions to colourize RC bug fixing packages differently
This commit adds some functions to query the BTS Soap interface to
determine if the package closes an RC bug or not. If it does close an
RC bug, it allows for differently colourizing the package line. It
adds a variable to indicate if a package fixes an RC bug or not, and
already allows for colourization of the HTML output; the new entry can
be used to colourize in other output formats as well.

Signed-off-by: Kumar Appaiah <a.kumar@alumni.iitm.ac.in>