rename volatile to updates
what an insanity. We end up with things like Squeeze-Updates-Proposed-Updates.
Why the heck one cant just make p-u a full suite and recommend people to use that
too? Its after a "NEW" queue anyways. This only serves to needlessly confuse
users. "What? proposed-updates and updates? whats going on?"
Meh. Volatile might be a bad name, but now this is worse.

Signed-off-by: Joerg Jaspert <joerg@debian.org>