Improve the label text and formatting of the labels for the queue graphs.
Rename the queue graph labels to ones shorter and of equal length.

Don't attempt to merge the numbers into the label, use cur instead.

Also use labels that are more indicative of their meaning.

Don't force the numbers to 3 digits in length.

Turn Min/Max/Avg in the labels to lower case.