control-suite: sort list of desired packages as intended
The list of desired packages contained a string of the form
"package-name version architecture".  Sorting with cmp_package_version
would sort this my the first two characters of the package name which is
not what was intended.

With this patch, the list instead contains tuples which means
cmp_package_version can now sort the list as intended.