queue_report: Ignore stuff without a source
Trainees can't see files in byhand, therefore process_changes_file in
queue_report is unable to handle changes files in queue/byhand. As a
workaround, whenever it encounters a changes file that it can't read
(and thus, cache[filename] has no source key), ignore that and
continue.

The result is that trainees will be able to use queue-report without a
backtrace, they just won't be able to see byhand, as intended. For
everyone else, queue-report should work as before.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>