debianqueued: set LC_ALL in the environment
Make sure all processes we start also run in the C locale.

queued currently parses the output from gpg to get the name of the
person who signed an upload. A better fix would be to pass --with-colon
to gpg, but as this information is only used for logging anyway...