"Update" Debian packaging
Whilst the packaging is obviously not for use, I'm finding it extremely
useful at least having the Build-Depends at a point where it will run the
tests.

It might be tempting to remove debian/ entirely, but Build-Depends is
just too useful for specifying, well, dependencies that tools such as
mk-build-deps can consume just like other packages.

Signed-off-by: Chris Lamb <lamby@debian.org>