generate-releases: Use the right key when testing uncompnotseen
This issue caused dak to not generate hashes for uncompressed Contents
files for some architectures in some components.  A bit of context,
there were up to 4 Contents files per architecture:

 * Contents-ARCH
 * main/Contents-ARCH
 * contrib/Contents-ARCH
 * non-free/Contents-ARCH

Any */Contents-ARCH that appeared on the filesystem AFTER the legacy
"Contents-ARCH" would be ignored, because generate-releases used the
wrong key to check if the file was already in uncompnotseen.

Signed-off-by: Niels Thykier <niels@thykier.net>