Pass suite name to the autobyhand scripts
To: debian-dak@lists.debian.org
Cc: Julien Cristau <jcristau@debian.org>
Date: Wed,  2 Dec 2015 08:01:37 +0100 (14 hours, 9 minutes, 8 seconds ago)
Resent-From: debian-dak@lists.debian.org

We have it readily available, and most scripts try to get it back from
the changes file anyway.

This makes things consistent for the scripts, so they won't get confused
by aliases (due to codenames or suite mappings) depending on whatever is
in the .changes file.

Signed-off-by: Julien Cristau <jcristau@debian.org>
Signed-off-by: Joerg Jaspert <joerg@debian.org>