Use maximum of modification and change time for detecting old files
When uploading using scp, the mtime might be set to the time from the
uploader's host.  Especially for the upstream tarballs this might result
in a mtime a long time in the past and queued might remove the uploaded
files right away as stray files.

Note that this removal only happens when the file is not processed: for
example when the upstream tarball gets uploaded (with an old mtime) and
queued processes uploads and removes stray files before the .changes for
the upload including the "old" tarball arrives.