Don't use dsync to de-dup files
The deduplication causes all files to get new mtimes in each round.
This forces rsync on each mirror to re-download the files and fix all
the hard-links.

Currently the archive shows about 4GiB of files in pool/ that have been
de-duplicated.  Most of the others are part of the installer or by-hash.