reject uploads that might cause conflicts with buildd uploads
For policy queues dak keeps the .buildinfo and .changes around until
the upload is processed.  As the uploads might already be sent to
buildds while waiting in the policy queue, the buildd upload might use
the same filename and end up getting rejected.  With this change, the
upload getting rejected should be the maintainer's upload.