manage-debug-suites: remove sourceless binaries too
For unstable dak adds the source package to the debug suite, but we
don't do that for testing (though dak will ensure the source is
present and keep it as an extra source in the Sources index).

This change makes sure to also remove obsolete debug packages that had
no source in the debug suite.