add rsync-staged script
This script is intended to be used when rsyncing uploads to another
host, for example from security-master.d.o to ssh.upload.d.o.

As the synced uploads include both binary-only and source uploads, we
have to make sure the source uploads are available before the
binary-only uploads are processed.  So we sync to a staging area and
and only move all files to the target directory afterwards.

There is still a small race condition as the source .changes might be
moved after the binary .changes, but dak should only reject uploads
after a short time (`Dinstall::SkipTime`).