Make dbtests pass and execute them as integration tests.
The dbtests had been left to rot, this change adapts them to the current state
of the database, which includes deleting code that was testing deleted code,
updating the code to the latest changes, etc.

Additionally, this change is also adding a run-dbtests script that will do
the full setup of the environment as expected by the tests, run the tests and
then clean up the environment.  This script is now executed from the
integration-tests/run-tests script, inside a postgres virtualenv.

Finally, this includes some additional changes to the setup scripts, to make it
possible to use them for database that don't use the default name.