dakweb: /sha256sum_in_archive/<sha256sum> endpoint
This patch creates a new /sha256sum_in_archive/<sha256sum>
endpoint in dakweb. The endpoint allows for searching for
files with a marching sha256sum in the archive.

I have also taken the time to fix docstrings of file_in_archive
which wrongly indicated that the return type was a dictionary.