Only require the .changes Binary and Description fields on binary uploads
For source-only uploads the fields should have been empty all along,
because there are no binary packages to include, but dpkg-genchanges
has incorrectly included all binary names that can be produced from
the debian/control template file (see #818618). Although DAK allows
an empty Binary field, the correct behavior should be for the field
to not be present if there are not binary packages to upload.