Stop requiring a Binary field in .changes files in the database
The changes table requires a not NULL binaries column. But we should not
be requiring a Binary field for source-only uploads. There have been
other changes to accommodate that, but the schema still requires the
values, and uploading such .changes files returns this misleading
rejection message:

  source_version_arch.changes is already known.

from daklib/archive.py:_install_changes() due to an exception when
inserting the DBChange() into the database.