dak admin: unify get_suite "or die" logic
At the moment, most "dak admin" subcommands implement the logic in the
form of:

  s = get_suite(s_n, ...)
  if s is None:
      die("...")

But the callers do not always agree on whether s_n to have a
".lower()" or not.

With this, change we unify this code flow so all of them now use the
new "get_suite_or_die", which will always "lower()" the suite name
plus handle this check for us.

A few calls that were missing the "if s is None: die()" code has been
migrated as well to fix some errors when the code operated on the None
object.

Signed-off-by: Niels Thykier <niels@thykier.net>