Process binaries is deterministic order
The order of the binaries shouldn't make a difference, but if it does, make
sure it is always the same.

As an example, in https://bugs.debian.org/824169, the source in the debug
archive gets the component from the first binary that is processed. When that
order is deterministic, the bug happens always or never, depending on whether
the component from the first binary is correct, instead of randomly on each
upload.