check binary archs for newer source cruft removal
Only remove sources that have been superseded by a newer version in a lower
suite if there are newer binaries for all relevant architectures.

This fixes 2 issues:

BinNMUs in testing-proposed-updates were immediately removed, because the
source with the same version was already in testing.

Sources in experimental were removed as soon as a newer source version was
accepted in unstable. For binaries only present in the experimental version,
this caused the overrides to be removed (as unused) in some cases. When the
binary upload for unstable came in, it would end up in binary-new.

With this change, the source in experimental will only be removed if unstable
has at least one newer binary for each architecture. This means the source
will only be removed from experimental after builds for all relevant
architectures have been accepted into unstable.