Remove Lintian tag 'source-field-does-not-match-pkg-name' from list of autorejects.
The Lintian tag cannot be produced using standard tools. It is
considered untestable in Lintian. Lintian no longer emits the tag.

The check was converted to a fatal program error. Users watching
Lintian's exit status and error output can probably still detect the
underlying condition.

When dpkg-source is used with the argument -DSource="some-other-name",
it will produce an error similar to:

    dpkg-source: error: source package has two conflicting values - some-other-name and correct-name

Removes the tag from list of FTP autorejects.