Fix unicode error in examine package
This error happens when there's a non-ascii character in Description
field, as it was the case for dpic_2019.08.30-1_amd64.changes, i.e.
there was a non-ascii dash character «‐» copied and pasted from the
description in upstream webpage.

```
$ env PYTHONPATH=. dak/examine_package.py dpic_2019.08.30-1_amd64.changes
Traceback (most recent call last):
  File "dak/examine_package.py", line 749, in <module>
    main()
  File "dak/examine_package.py", line 710, in main
    check_changes(f)
  File "dak/examine_package.py", line 663, in check_changes
    print(display_changes(changes['distribution'], changes_filename))
UnicodeEncodeError: 'ascii' codec can't encode character u'\u2010' in position 332: ordinal not in range(128)
```

Signed-off-by: Alex Muntada <alexm@debian.org>