control-suite: allow equal versions when doing version checks
Accept equal version for MustBeNewerThan and MustBeOlderThan in control-suite.

When using control-suite, it can be useful to accept the same version when
doing version comparison.

When a binNMU is uploaded to testing-proposed-updates, dak copies the source
from testing to t-p-u even though the source version is the same. When adding
a binNMU to t-p-u with control-suite, however, the strict 'MustBeNewerThan
testing' doesn't allow the source to be added to t-p-u as well.

Adding a binary to t-p-u without the source leads to different behaviour than
uploading it. Moreover, a binary in t-p-u without the corresponding source
confuses wanna-build and britney.

This change will allow control-suite to add equal versions even if a
MustBeNewerThan or MustBeOlderThan version check is specified.

This change doesn't change the version check behaviour for uploads. However,
uploads of packages identical to a version already in the archive are usually
not wanted.