dak rm: make --outdated an option instead of a removal mode
Allow the --outdated option to be used with all removal modes. No longer
consider it to be a separate removal mode.

Note that this slightly changes the behaviour when using --outdated without
specifying a removal mode. The new behaviour also removes the outdated
sources. In practice, this won't matter a lot, because those outdated sources
would be removed automatically after the removal of the binaries.

The old behaviour can be restored by using --binary-only together with
--outdated.