control-suite: set extra_suites to target suite's archive
When copying to a suite with an associcated build queue
`control-suite` will copy to the build queue as well.  This
would fail if an added binary package referred to a source
package via `Built-Using` and said source package was not
in the build queue's archive.

We now allow copying the source package in question to the
build queue's archive automatically if it is already present
in the target suite's archive.