Reduce the number of garbage collected file descriptors
The code base contains a few unfornuate patterns such as:

    data = open(...).read()

OR

    fh = open(...)
    data = fh.read()

This causes a file descriptor leak as the opened file is never
explicitly closed again.  While slightly mitigated by CPython using
ref-counting and immediate closure, we should still update the code
to close the files.

This commit focuses on a large number of simple clean ups.  Most of
which relies on the "with open(...) as X" pattern where feasible.

As "with" causes identation noise, this commit deliberately refains
from "fixing" other suboptimal behaviour.  Notably there are several
instances of "except:" visible in the diff that should ideally be
reduced.

Signed-off-by: Niels Thykier <niels@thykier.net>