For FTP auto rejection of non UTF-8 filenames, add a renamed Lintian tag but retain the old one.
Adds a renamed Lintian tag to dak's FTP auto-reject profile. Retains
the old tag until the transition is complete.

Here is more information about the changes in Lintian:

To resolve Bug#956723, Lintian recently introduced two new tags that
relate to files whose names are not valid UTF-8, and presumably use a
national encoding:

    upstream-file-without-utf8-name (files in orig.tar.gz)
    patched-file-without-utf8-name (files created in Debian patches)

You can find the relevant commit here:

    https://salsa.debian.org/lintian/lintian/-/commit/bc81771b388cbf53cf7ecb0e530108212d891565

There already was a corresponding tag for installation packages, but
it was named very broadly. That tag was renamed and is now similar to
the other two:

    shipped-file-without-utf8-name (files in .deb or .udeb packages)

The commit is here:

    https://salsa.debian.org/lintian/lintian/-/commit/b2fbc44f68473f7e2e74857d3a4c623d921ccdc7

This merge request was not tested in any way.